Binary Search
Binary search is a fast way to find something in a sorted list.
In each step, this algorithm divides the list into 2 parts.

If the target is at the middle index, this is the best case, and the time complexity is O(1).

If the target is near the beginning or end, this is the worst case, and the time complexity is O(log n).

The average case is also O(log n).

Binary search is faster than linear search because it reduces the search range by half every time.